+++ minigalaxy/ui/window.py
@@ -2,6 +2,8 @@ import os
 import locale
 
 from minigalaxy.download_manager import DownloadManager
+from minigalaxy.logger import logger
+from minigalaxy.ui.categoryfilters import CategoryFilters
 from minigalaxy.ui.login import Login
 from minigalaxy.ui.preferences import Preferences
 from minigalaxy.ui.about import About
@@ -76,8 +78,8 @@ class Window(Gtk.ApplicationWindow):
             try:
                 self.__authenticate()
                 self.HeaderBar.set_subtitle(self.api.get_user_info())
-            except Exception as e:
-                print("Starting in offline mode, after receiving exception: {}".format(e))
+            except Exception:
+                logger.warn("Starting in offline mode after receiving exception", exc_info=1)
                 self.offline = True
         self.sync_library()
 
@@ -98,6 +100,12 @@ class Window(Gtk.ApplicationWindow):
         about_window = About(self)
         about_window.run()
         about_window.destroy()
+
+    @Gtk.Template.Callback("on_menu_category_filter_clicked")
+    def show_categories(self, button):
+        category_filters_window = CategoryFilters(self, self.library)
+        category_filters_window.run()
+        category_filters_window.destroy()
 
     @Gtk.Template.Callback("on_menu_logout_clicked")
     def logout(self, button):