+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added additional tooltips to buttons, labels, menu items and radio buttons (thanks to orende)
 - Hide CDPR Goodie Pack Content
 - Add notifications on successful download and installation of games (thanks to orende)
+- Add category filtering dialog for game library (thanks to orende)
 
 **1.2.2**
 - Fix progress bar not showing up for downloads